requirements.txt
# Core data science libraries
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Statistical analysis
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Environment management
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Optional: Database connectivity (if needed for real GNPC data)
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0
